refactor(mistral3.5): drop Devstral path, rename module to mistral3_vlm_fp8
The FP8 custom-model module was originally added to cover Devstral (24B
and 123B text-only checkpoints) alongside the Mistral-3.5 128B VLM
(dawn-ridge). We no longer need the Devstral variants, so keep only the
VLM class and rename the module to match its actual scope.

Removed:
- Mistral3FP8ForCausalLM (text-only wrapper around Ministral3ForCausalLM)
- Mistral3FP8StateDictAdapter factories: for_dense, for_devstral_vlm,
  for_dawn_ridge_vlm (and their key-rewrite helpers)
- _detect_layout / _resolve_snapshot_dir (checkpoint-layout sniffing)
- Text-only branch of the resolver hook (the VLM branch remains)
- examples/llm_finetune/devstral/devstral2_* YAMLs
- examples/llm_finetune/devstral/mistral3p5_128b_hellaswag_tp_pp.yaml
  (text-only variant of the 128B recipe)
- devstral_123b.sub, devstral_24b_tp_pp_2node.sub, mistral3p5_128b.sub
- Devstral side-effect import in recipes/llm/train_ft.py

Kept:
- Mistral3FP8VLMForConditionalGeneration and for_vlm_full adapter
- Resolver hook dispatching VLM via NeMoAutoModelForImageTextToText
- MedPix recipe + sbatch (mistral3p5_128b_medpix.sub) — that's the live
  training path with verified step-0 loss 3.20 (HF ref 3.47).

All PP/TP wiring lands on the new module path
``nemo_automodel.components.models.mistral3_vlm_fp8``; the registered
TP-plan qualname is updated to match.
