--- devstral_24b_tp_pp_2node.sub
@@ -1,56 +1,0 @@
-#!/bin/bash
-
-# =============================================================================
-# Tier-1 smoke: Devstral-Small-2-24B-Instruct-2512 full-FT SQuAD on
-# 2 nodes x 8 H100-80GB = 16 GPUs, TP=2 PP=2 (DP=4).
-# Gates the 123B multi-node run.
-#
-# Uses the custom streaming FP8 loader (registered for
-# Mistral3ForConditionalGeneration in nemo_automodel/components/models/devstral/),
-# which avoids HF's per-rank full-BF16 materialization under PP. Per-rank peak
-# stays under ~35 GB on 80 GB H100. Verified end-to-end on a single 8xH100 node.
-# =============================================================================
-
-# Parameters
-#SBATCH --account=coreai_dlalgo_genai
-#SBATCH --dependency=singleton
-#SBATCH --exclusive
-#SBATCH --job-name=coreai_dlalgo_genai-devstral_24b_tp2pp2
-#SBATCH --mem=0
-#SBATCH --nodes=2
-#SBATCH --gpus-per-node=8
-#SBATCH --ntasks-per-node=8
-#SBATCH --time=00:30:00
-#SBATCH --output=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel/slurm/devstral_24b_tp2pp2_%j.out
-#SBATCH --error=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel/slurm/devstral_24b_tp2pp2_%j.err
-
-# setup
-export TRANSFORMERS_OFFLINE=1
-export HF_DATASETS_OFFLINE=1
-export HF_HUB_OFFLINE=1
-export HF_HOME=/lustre/fsw/portfolios/coreai/users/huiyingl/hf_cache
-export TORCH_NCCL_AVOID_RECORD_STREAMS=1
-export NCCL_NVLS_ENABLE=0
-export TORCH_NCCL_TRACE_BUFFER_SIZE=2000
-export TORCH_NCCL_DUMP_ON_TIMEOUT=1
-export WANDB_API_KEY=70ce214ec4036cffa29729bf4dbc2e299eac4921
-
-srun --container-image /lustre/fsw/portfolios/coreai/users/huiyingl/container_images/48758702.sqsh --container-mounts /lustre/:/lustre/ --no-container-mount-home bash -c "
-  cd /lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel;
-  export PYTHONPATH=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel;
-  export WANDB_API_KEY=70ce214ec4036cffa29729bf4dbc2e299eac4921;
-  export TRANSFORMERS_OFFLINE=1;
-  export HF_DATASETS_OFFLINE=1;
-  export HF_HUB_OFFLINE=1;
-  export HF_HOME=/lustre/fsw/portfolios/coreai/users/huiyingl/hf_cache;
-  export TORCH_NCCL_TRACE_BUFFER_SIZE=2000;
-  export TORCH_NCCL_DUMP_ON_TIMEOUT=1;
-  export TORCH_FR_DUMP_TEMP_FILE=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel/slurm/nccl/nccl_dump_%j.json;
-  export AUTOMODEL_DEBUG_MEM=1;
-  CUDA_DEVICE_MAX_CONNECTIONS=1 CUDA_VISIBLE_DEVICES=0,1,2,3,4,5,6,7 python3 examples/llm_finetune/finetune.py \
-  -c examples/llm_finetune/devstral/devstral2_small_2512_squad_tp2pp2.yaml \
-  --step_scheduler.max_steps=50 \
-  --wandb.project=huiyingl_workspace \
-  --wandb.name=devstral_24b_tp2pp2_2node \
-  --wandb.entity=Nemo-automodel \
-  --checkpoint.enabled=false "
