--- examples/llm_finetune/devstral/devstral2_123b_hellaswag_tp_pp.yaml
@@ -1,121 +1,0 @@
-# Tier-2: Devstral-2-123B-Instruct-2512 on HellaSwag, full FT.
-# Multi-node: 8 nodes x 8 H100-80GB = 64 GPUs, TP=8 PP=8, DP=1.
-#
-# FP8 load path:
-#   Devstral-2-123B-Instruct-2512 ships as FP8 safetensors
-#   (arch=Ministral3ForCausalLM, quant_method=fp8, per-tensor scalar scale).
-#   HF's FineGrainedFP8Config loader materializes the full dequantized BF16
-#   model on every rank before PP split, which would OOM a 123B model on any
-#   reasonable GPU.
-#
-#   We instead use the DSV3-style state_dict_adapter pattern:
-#     1. nemo_automodel/components/models/devstral/__init__.py installs a
-#        resolver hook so FP8-native Ministral3 configs dispatch to
-#        Devstral123BFP8ForCausalLM.
-#     2. That class constructs an empty Ministral3 text decoder on meta and
-#        attaches a DevstralFP8StateDictAdapter(key_prefix="") — 123B has no
-#        language_model. prefix on disk, just `model.*` + `lm_head`.
-#     3. Standard Checkpointer.load_model → DCP fetches weight + weight_scale_inv
-#        per PP stage; adapter.from_hf pairs them, dequantizes to BF16, and
-#        drops the scale/activation_scale keys. set_model_state_dict
-#        distributes each rank's local DTensor shard.
-#
-# Launch:
-#   sbatch /lustre/fsw/portfolios/coreai/users/huiyingl/dev123/Automodel/devstral_123b.sub
-# Only submit after tier-1 (24B TP=2 PP=2, devstral2_small_2512_squad_tp2pp2.yaml)
-# has passed — gate on a loss-decreasing two-step run on a single node.
-
-recipe: TrainFinetuneRecipeForNextTokenPrediction
-
-step_scheduler:
-  global_batch_size: 256
-  # Must be >= pp_size * pp_microbatch_size = 8 * 1 = 8. Each grad-accumulation
-  # step submits local_batch_size / pp_microbatch_size = 8 microbatches into
-  # the 1f1b pipeline, which is the minimum needed to fill all 8 PP stages.
-  # DP=1, so grad_acc = global_batch_size / local_batch_size = 256 / 8 = 32.
-  local_batch_size: 8
-  ckpt_every_steps: 1000
-  val_every_steps: 20
-  max_steps: 50
-
-dist_env:
-  backend: nccl
-  timeout_minutes: 10
-
-rng:
-  _target_: nemo_automodel.components.training.rng.StatefulRNG
-  seed: 1111
-  ranked: true
-
-model:
-  # No `quantization_config` — our registered Devstral123BFP8ForCausalLM
-  # handles FP8 dequant via its state_dict_adapter. HF's FineGrainedFP8
-  # loader is never invoked.
-  _target_: nemo_automodel.NeMoAutoModelForCausalLM.from_pretrained
-  pretrained_model_name_or_path: mistralai/Devstral-2-123B-Instruct-2512
-
-checkpoint:
-  enabled: false
-  checkpoint_dir: logs/tier2_devstral_123b_hellaswag/ckpt
-  model_save_format: safetensors
-  save_consolidated: false
-
-distributed:
-  strategy: fsdp2
-  dp_size: none
-  tp_size: 8
-  cp_size: 1
-  pp_size: 8
-  # sequence_parallel MUST be false when combining TP with PP — SP shards
-  # activations on the sequence dim as DTensors, but PyTorch's 1f1b schedule
-  # sends activations between PP stages via plain c10d::send, which fails
-  # with "found no DeviceMesh from dtensor args for c10d::send".
-  # Matches the qwen3.5_27b_tp4pp4 / gemma4_31b_tp4_pp2 references.
-  sequence_parallel: false
-  activation_checkpointing: true
-  defer_fsdp_grad_sync: false
-  pipeline:
-    # Non-interleaved 1f1b with auto layer split. Ministral3 has 88 layers →
-    # 88/8 = 11 layers per stage.
-    pp_schedule: 1f1b
-    pp_microbatch_size: 1
-    scale_grads_in_schedule: false
-
-loss_fn:
-  _target_: nemo_automodel.components.loss.masked_ce.MaskedCrossEntropy
-
-dataset:
-  _target_: nemo_automodel.components.datasets.llm.hellaswag.HellaSwag
-  path_or_dataset: rowan/hellaswag
-  split: train
-
-packed_sequence:
-  packed_sequence_size: 0
-
-dataloader:
-  _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn: nemo_automodel.components.datasets.utils.default_collater
-  shuffle: true
-
-validation_dataset:
-  _target_: nemo_automodel.components.datasets.llm.hellaswag.HellaSwag
-  path_or_dataset: rowan/hellaswag
-  split: validation
-  num_samples_limit: 64
-
-validation_dataloader:
-  _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn:
-    _target_: nemo_automodel.components.datasets.utils.default_collater
-    pad_seq_len_divisible: 160
-
-optimizer:
-  _target_: torch.optim.Adam
-  betas: [0.9, 0.999]
-  eps: 1e-8
-  lr: 1.0e-5
-  weight_decay: 0
-
-lr_scheduler:
-  lr_decay_style: cosine
-  min_lr: 1.0e-6
