--- examples/llm_finetune/devstral/devstral2_small_2512_squad_peft.yaml
@@ -1,117 +1,0 @@
-# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-
-# To run this recipe:
-#   automodel examples/llm_finetune/devstral/devstral2_small_2512_squad_peft.yaml --nproc-per-node 8
-# Adjust --nproc-per-node to the number of GPUs available on your machine.
-
-recipe: TrainFinetuneRecipeForNextTokenPrediction
-
-step_scheduler:
-  global_batch_size: 64
-  local_batch_size: 1
-  ckpt_every_steps: 200
-  val_every_steps: 100  # will run every x number of gradient steps
-  num_epochs: 1
-
-dist_env:
-  backend: nccl
-  timeout_minutes: 1
-
-rng:
-  _target_: nemo_automodel.components.training.rng.StatefulRNG
-  seed: 1111
-  ranked: true
-
-model:
-  _target_: nemo_automodel.NeMoAutoModelForCausalLM.from_pretrained
-  pretrained_model_name_or_path: mistralai/Devstral-Small-2-24B-Instruct-2512
-  quantization_config:
-    _target_: transformers.FineGrainedFP8Config
-    dequantize: true
-
-peft:
-  _target_: nemo_automodel.components._peft.lora.PeftConfig
-  match_all_linear: True
-  dim: 8
-  alpha: 32
-  use_triton: True
-  # dtype needs a fix to resolve to type instead of string
-  # lora_dtype: torch.bfloat16
-
-checkpoint:
-  enabled: true
-  checkpoint_dir: checkpoints/
-  model_save_format: safetensors
-  save_consolidated: false # saves the model in a consolidated safetensors format. Requires model_save_format to be safetensors.
-
-distributed:
-  strategy: fsdp2
-  dp_size: none
-  tp_size: 1
-  cp_size: 1
-
-  sequence_parallel: false
-
-loss_fn:
-  _target_: nemo_automodel.components.loss.masked_ce.MaskedCrossEntropy
-
-dataset:
-  _target_: nemo_automodel.components.datasets.llm.squad.make_squad_dataset
-  dataset_name: rajpurkar/squad
-  split: train
-
-packed_sequence:
-  # Set packed_sequence_size > 0 to run with packed sequences
-  packed_sequence_size: 0
-
-dataloader:
-  _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn: nemo_automodel.components.datasets.utils.default_collater
-  shuffle: true
-
-
-validation_dataset:
-  _target_: nemo_automodel.components.datasets.llm.squad.make_squad_dataset
-  dataset_name: rajpurkar/squad
-  split: validation
-  limit_dataset_samples: 64
-
-validation_dataloader:
-  _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn: nemo_automodel.components.datasets.utils.default_collater
-
-optimizer:
-  _target_: torch.optim.Adam
-  betas: [0.9, 0.999]
-  eps: 1e-8
-  lr: 1.0e-5
-  weight_decay: 0
-  # min_lr: 1.0e-5
-
-lr_scheduler:
-  lr_decay_style: cosine
-  min_lr: 1.0e-6
-
-ci:
-  recipe_owner: akoumpa
-  time: "00:30:00"
-
-# Uncomment and configure for W&B logging
-# wandb:
-#   project: <your_wandb_project>
-#   entity: <your_wandb_entity>
-#   name: <your_wandb_exp_name>
-#   save_dir: <your_wandb_save_dir>
