--- examples/llm_finetune/devstral/devstral2_small_2512_squad_tp2pp2.yaml
@@ -1,118 +1,0 @@
-# Full-FT SQuAD on 24B Devstral-Small with TP=2 PP=2.
-# Fits on a single 8xH100-80GB node; scales out to 2 nodes (16 GPUs, DP=4).
-#
-# FP8 load path:
-#   Devstral-Small-2-24B-Instruct-2512 ships as FP8 safetensors
-#   (quant_method=fp8, per-tensor scalar scale). HF's FineGrainedFP8Config
-#   loader materializes the full dequantized BF16 model on each rank before
-#   PP split — that OOMs at ~77 GB/rank on 80 GB H100.
-#
-#   Our approach follows the same state_dict_adapter pattern as
-#   deepseek_v3/state_dict_adapter.py:
-#     1. nemo_automodel/components/models/devstral/__init__.py installs a
-#        resolver hook so FP8-native Mistral3 / Ministral3 configs dispatch to
-#        our custom classes. Non-FP8 models are untouched (no registry override).
-#     2. The custom class constructs a Ministral3 text decoder on meta and
-#        attaches a DevstralFP8StateDictAdapter.
-#     3. Standard Checkpointer.load_model -> DCP fetches weight + weight_scale_inv
-#        per PP stage (the adapter's to_hf adds FP8-typed placeholders so DCP
-#        reads FP8 bytes verbatim from safetensors).
-#     4. adapter.from_hf pairs each FP8 weight with its scalar scale_inv, emits
-#        bf16, drops the scale/activation_scale keys; set_model_state_dict
-#        distributes each rank's local DTensor shard.
-#
-#   Per-rank peak stays under ~35 GB; verified end-to-end on 1 node 8xH100:
-#     step 0: loss 0.82, grad_norm 149
-#     step 1: loss 0.50, grad_norm  87
-#
-# Launch via devstral_24b_tp_pp_2node.sub for multi-node runs.
-
-recipe: TrainFinetuneRecipeForNextTokenPrediction
-
-step_scheduler:
-  global_batch_size: 16
-  local_batch_size: 2        # must be >= pp_size * pp_microbatch_size = 2
-  ckpt_every_steps: 1000
-  val_every_steps: 1000
-  num_epochs: 1
-  max_steps: 50
-
-dist_env:
-  backend: nccl
-  timeout_minutes: 5
-
-rng:
-  _target_: nemo_automodel.components.training.rng.StatefulRNG
-  seed: 1111
-  ranked: true
-
-model:
-  # Standard NeMoAutoModel entry point. A resolver hook in
-  # `nemo_automodel/components/models/devstral/__init__.py` claims FP8-native
-  # Mistral3 configs for Devstral24BFP8TextForCausalLM — the class builds a
-  # text-only Ministral3 decoder on meta and attaches DevstralFP8StateDictAdapter,
-  # which the standard Checkpointer load path (_maybe_adapt_state_dict_{to,from}_hf)
-  # calls to dequantize FP8 → BF16. No `quantization_config` here: if one were
-  # passed, `_init_model` would fall through to HF's FineGrainedFP8 loader
-  # (the thing that OOMs under PP).
-  _target_: nemo_automodel.NeMoAutoModelForCausalLM.from_pretrained
-  pretrained_model_name_or_path: mistralai/Devstral-Small-2-24B-Instruct-2512
-
-checkpoint:
-  enabled: false
-
-distributed:
-  strategy: fsdp2
-  dp_size: none
-  tp_size: 2
-  cp_size: 1
-  pp_size: 2
-  sequence_parallel: false
-  activation_checkpointing: true
-  defer_fsdp_grad_sync: false
-  pipeline:
-    # Match the reference HF-model configs (gemma4_31b_tp4_pp2, qwen3.5_27b_tp4pp4):
-    # non-interleaved 1f1b schedule with auto layer split (40 layers / pp_size).
-    # patch_inner_model and patch_causal_lm_model default to True — correct for
-    # Mistral3ForConditionalGeneration (has .model.language_model; patches both
-    # the inner Ministral3 text backbone and the outer CausalLM wrapper).
-    pp_schedule: 1f1b
-    pp_microbatch_size: 1
-    scale_grads_in_schedule: false
-
-loss_fn:
-  _target_: nemo_automodel.components.loss.masked_ce.MaskedCrossEntropy
-
-dataset:
-  _target_: nemo_automodel.components.datasets.llm.squad.make_squad_dataset
-  dataset_name: rajpurkar/squad
-  split: train
-
-packed_sequence:
-  packed_sequence_size: 0
-
-dataloader:
-  _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn: nemo_automodel.components.datasets.utils.default_collater
-  shuffle: true
-
-validation_dataset:
-  _target_: nemo_automodel.components.datasets.llm.squad.make_squad_dataset
-  dataset_name: rajpurkar/squad
-  split: validation
-  limit_dataset_samples: 64
-
-validation_dataloader:
-  _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn: nemo_automodel.components.datasets.utils.default_collater
-
-optimizer:
-  _target_: torch.optim.Adam
-  betas: [0.9, 0.999]
-  eps: 1e-8
-  lr: 1.0e-5
-  weight_decay: 0
-
-lr_scheduler:
-  lr_decay_style: cosine
-  min_lr: 1.0e-6
