--- mistral3p5_128b.sub
@@ -1,56 +1,0 @@
-#!/bin/bash
-
-# =============================================================================
-# Mistral-3.5 medium 128B (codename "dawn-ridge") full-FT HellaSwag on
-# 8 nodes x 8 H100-80GB = 64 GPUs, TP=8 PP=8 (DP=1).
-#
-# Uses the shared FP8 state-dict adapter (Mistral3FP8ForCausalLM +
-# Mistral3FP8StateDictAdapter in nemo_automodel/components/models/devstral/)
-# which auto-detects the dawn-ridge VLM key layout (`model.language_model.X`
-# infix + top-level `lm_head`). No HF FineGrainedFP8 allocation path is
-# invoked.
-# =============================================================================
-
-# Parameters
-#SBATCH --account=coreai_dlalgo_genai
-#SBATCH --dependency=singleton
-#SBATCH --exclusive
-#SBATCH --job-name=coreai_dlalgo_genai-mistral3p5_128b_tp8pp8
-#SBATCH --mem=0
-#SBATCH --nodes=8
-#SBATCH --gpus-per-node=8
-#SBATCH --ntasks-per-node=8
-#SBATCH --time=00:30:00
-#SBATCH --output=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel/slurm/mistral3p5_128b_tp8pp8_%j.out
-#SBATCH --error=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel/slurm/mistral3p5_128b_tp8pp8_%j.err
-
-# setup
-export TRANSFORMERS_OFFLINE=1
-export HF_DATASETS_OFFLINE=1
-export HF_HUB_OFFLINE=1
-export HF_HOME=/lustre/fsw/portfolios/coreai/users/huiyingl/hf_cache
-export TORCH_NCCL_AVOID_RECORD_STREAMS=1
-export NCCL_NVLS_ENABLE=0
-export TORCH_FR_BUFFER_SIZE=2000
-export TORCH_NCCL_DUMP_ON_TIMEOUT=1
-export WANDB_API_KEY=70ce214ec4036cffa29729bf4dbc2e299eac4921
-
-srun --container-image /lustre/fsw/portfolios/coreai/users/huiyingl/container_images/48758702.sqsh --container-mounts /lustre/:/lustre/ --no-container-mount-home bash -c "
-  cd /lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel;
-  export PYTHONPATH=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel;
-  export WANDB_API_KEY=70ce214ec4036cffa29729bf4dbc2e299eac4921;
-  export TRANSFORMERS_OFFLINE=1;
-  export HF_DATASETS_OFFLINE=1;
-  export HF_HUB_OFFLINE=1;
-  export HF_HOME=/lustre/fsw/portfolios/coreai/users/huiyingl/hf_cache;
-  export TORCH_FR_BUFFER_SIZE=2000;
-  export TORCH_NCCL_DUMP_ON_TIMEOUT=1;
-  export TORCH_FR_DUMP_TEMP_FILE=/lustre/fs1/portfolios/coreai/projects/coreai_dlalgo_nemofw/users/huiyingl/dev123/Automodel/slurm/nccl/nccl_dump_%j.json;
-  export AUTOMODEL_DEBUG_MEM=1;
-  CUDA_DEVICE_MAX_CONNECTIONS=1 CUDA_VISIBLE_DEVICES=0,1,2,3,4,5,6,7 python3 examples/llm_finetune/finetune.py \
-  -c examples/llm_finetune/devstral/mistral3p5_128b_hellaswag_tp_pp.yaml \
-  --step_scheduler.max_steps=50 \
-  --wandb.project=huiyingl_workspace \
-  --wandb.name=mistral3p5_128b_tp8pp8_8node \
-  --wandb.entity=Nemo-automodel \
-  --checkpoint.enabled=false "
